Move add_covariate_effect tests to own file
